+++ src/component/emoji-card/EmojiCard.jsx
