+++ src/component/navbar/Navbar.jsx
@@ -1,7 +1,7 @@
 import { useContext } from "react";
 import "./Navbar.css";
-import { AuthContext } from "../context/AuthContext";
-import { Search } from "./Search";
+import { AuthContext } from "../../context/AuthContext";
+import { Search } from "../search/Search";
 import { useNavigate } from "react-router";
 export const Navbar = () => {
   const { user } = useContext(AuthContext);
@@ -12,17 +12,10 @@ export const Navbar = () => {
       <nav className="nav-container">
         <img
           className="nav-brand-logo"
-          src={require("../images/logo.png")}
+          src={require("../../images/logo.png")}
           alt="hangout logo"
         />
-
-        {/* <input
-          type="search"
-          className="search-input"
-          placeholder="Search User"
-        /> */}
         <Search />
-
         <img
           className="nav-user-image"
           src={profileImg}
