+++ src/component/post/Post.jsx
@@ -6,13 +6,13 @@ import BookmarkIcon from "@mui/icons-material/Bookmark";
 import ShareIcon from "@mui/icons-material/Share";
 import "./Post.css";
 import { useContext, useState } from "react";
-import { PostContext } from "../context/PostContext";
-import { AuthContext } from "../context/AuthContext";
-import { UserContext } from "../context/UserContext";
+import { PostContext } from "../../context/PostContext";
+import { AuthContext } from "../../context/AuthContext";
+import { UserContext } from "../../context/UserContext";
 import { useNavigate } from "react-router";
 import { Popper } from "@mui/material";
-import { Modal } from "./Modal";
-import { EditPost } from "./EditPost";
+import { Modal } from "../modal/Modal";
+import { EditPost } from "../edit-post/EditPost";
 export const Post = ({ postDetails }) => {
   const { _id, content, createdAt, likes, username, mediaURL } = postDetails;
   const { toggleLikeHandler, isLikedHandler, deletePost } =
