+++ src/component/suggestion-card/SuggestionBar.css
